requirements.txt
absl-py==1.4.0
albumentations==1.4.0
antlr4-python3-runtime==4.8
anyio==3.6.2
anykeystore==0.2
apex==0.9.10.dev0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.2
asttokens==2.2.1
asynctest==0.13.0
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
cryptacular==1.6.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
et-xmlfile==1.1.0
executing==1.2.0
faiss-gpu==1.6.5
fastjsonschema==2.16.3
filelock==3.12.4
flake8==6.0.0
flit_core==3.8.0
fonttools==4.39.3
fqdn==1.5.1
fsspec==2023.10.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.17.3
google-auth-oauthlib==0.4.6
greenlet==3.1.1
grpcio==1.53.0
haversine==2.8.0
huggingface-hub==0.17.3
hupper==1.12.1
hydra-colorlog==1.1.0
hydra-core==1.1.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipdb==0.13.13
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.3
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
packaging==23.1
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
PasteDeploy==3.1.0
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.0
pylint==2.17.2
pyparsing==3.0.9
pyramid==2.0.2
pyramid-mailer==0.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytorch-metric-learning==1.5.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
qudida==0.0.4
regex==2023.10.3
repoze.sendmail==4.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.4.0
SciencePlots==2.0.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.37
stack-data==0.6.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.8
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
tomlkit==0.11.7
torch==1.8.1+cu102
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1+cu102
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
transaction==5.0
transformers==4.34.1
translationstring==1.4
typing_extensions==4.13.2
uri-template==1.2.0
urllib3==1.26.15
velruse==1.1.1
venusian==3.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.5.1
Werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==4.0.7
wrapt==1.15.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
zipp==3.15.0
zope.deprecation==5.0
zope.interface==7.2
zope.sqlalchemy==3.1